third_party/WebKit/LayoutTests/fast/media/mq-hover.html

<html>
<head>
<title>Test the CSS4 media query "hover" features.</title>
<style type="text/css">

</style>
<script type="text/javascript" charset="utf-8">
    if (window.testRunner)
        testRunner.dumpAsText();

    function log(m) {
        document.getElementById('results').innerHTML += m + '<br>';
    }

    function testQueries()
    {
        var queries = [
            "(hover)",
            "(Hover)",
            "(hover:none)",
            "(hover:NoNe)",
            "(hover:on-demand)",
            "(hover:bogusvalue)",
            "(hover:hover)",
            "(hover:HoveR)",
            "(any-hover)",
            "(any-Hover)",
            "(any-hover:none)",
            "(any-hover:NoNe)",
            "(any-hover:on-demand)",
            "(any-hover:bogusvalue)",
            "(any-hover:hover)",
            "(any-hover:HoveR)"
            ];
        for (var index=0; index < queries.length; index++) {
            log("Query &quot;" + queries[index] + "&quot: " +
                window.matchMedia(queries[index]).matches);
        }
    }

    function runTests()
    {
        if (!window.internals)
            return;

        log("------------- with 'hover' hover -------------");
        internals.settings.setPrimaryHoverType('hover');
        internals.settings.setAvailableHoverTypes('hover');
        testQueries();

        log("------------- with 'none' and 'hover' hover, 'none' is primary -------------");
        internals.settings.setPrimaryHoverType('none');
        internals.settings.setAvailableHoverTypes('none,hover');
        testQueries();

        log("------------- with 'none' hover -------------");
        internals.settings.setPrimaryHoverType('none');
        internals.settings.setAvailableHoverTypes('none');
        testQueries();
    }

</script>
</head>
<body onload="runTests()">
    <p>Test the <a href="http://www.w3.org/TR/2014/WD-mediaqueries-4-20140605/#hover">(hover)</a> media features.
    See <a href="http://crbug.com/136119">Bug 136119</a> for details.</p>

    <div id="results">
    </div>
</body>
</html>
